refactor: utils 파일 분리 및 print_log 로직 수정
